CSN: CSN-multifpga-61
Use SCons "Copy" factory instead of "cp" because SCons stores the signature of
the program in the build dependence.  If "cp" is a different binary it looks like
rebuilding is needed.

Removed mkdir from the rules instead of using "Mkdir" factory, since the
directories appear to be present already and the factory version fails if
the target exists.