CSN: CSN-multifpga-103
Fixed some obvious and glaring bugs in LJF - namely we had some
holdovers from the old days when link 0 of a lane was always
flowcontrol.  As a result, whatever link was assigned to that lane had
no flow control run on it, leading to the occasional droppin of
data. This is fixed now. 

LJF gives a 2.75x speedup on the ACP over the Random, single lane for
the loop test.  I feel like this number should be closer to 5x though,
so there's probably a performance bug in the allocation algorithm.