CSN: CSN-multifpga-118
Added support for back-to-back reads from the same channel in the
ingress switch.  This required adding more precise buffer control
mechanims to the packetizers and soft-connections.

Also, I made the type generation for the lanes a little more precise.
Now we use the maximum width link actually allocated to the lane,
rather than the PHYSICAL_CONNECTION_SIZE.  This makes packets more
efficient. It also allows us to eliminate marshalling logic in some of
the packetizers.