CSN: CSN-multifpga-141
Fixed bug in load balancer code.  We made the assumption that header
were 12 bits wide, but this was causing load balancing mistakes for
narrow lanes. Now the load balancing is recursive. We set provisional
header widths, do a lane allocation, and check if the provisional
header sizes were feasible.  If they weren't, we store the correct
header sizes and iterate until a fixed point is reached.  Usually this
happens after a couple of iterations.  

Balancing errors in HAsim were as high as 40%, but now seem to be
under 10%.