CSN: CSN-multifpga-155
Timing closure was too difficult for egress arbiters that have a large number of input
ports.  The arbiter is now configurable for either the previous single cycle mode
or a new dual cycle mode in which an input channel is picked in one cycle and data
movement begins in the next cycle.  The dual cycle mode is pipelined except that the
same input channel may not be chosen in consecutive cycles.  Given that the arbiter
is already managing lots of input channels and consecutive cycles matter only when
the input channel's message size is only a single UMF chunk, this rarely affects
performance.

The current code in multiFPGAConnect.py choses multi cycle arbitration when more than
10 input ports are competing for a physical channel.