CSN: CSN-multifpga-216
Fixed bug in LI compiler.  To simplify software, we use fixed umf
protocol parameters as provided by the legacy umf code.  However,
these parameters were causing proviso failures in the router code
because some of the code was written in terms of the maximum value of
the parameter rather than the maximum value actually used in the
router. 

Changed the router to use the values provided by the actual channels,
rather than the absolute maximum of the parameter. Values that are too
large will still result in a compile time error, but only a
pathalogical program would run in to this case (i.e. LI channel width
>> 100000 bits).